# backend/gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API документация для budget-app",
        "title": "Budget API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/ledger/budgets/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ledger"
                ],
                "summary": "Get budget",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ledger.BudgetGetHandlerOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJSON"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "ledger"
                ],
                "summary": "Delete budget",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJSON"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ledger"
                ],
                "summary": "Patch budget",
                "parameters": [
                    {
                        "description": "JSON",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ledger.BudgetPatchHandlerInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ledger.BudgetPatchHandlerOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJSON"
                        }
                    }
                }
            }
        },
...
        "/ledger/reports": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ledger"
                ],
                "summary": "Reports",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по дате ОТ в формате 2025-01-30 (год-месяц-день)",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по дате ДО в формате 2025-01-30 (год-месяц-день)",
                        "name": "date_to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ledger.ReportListHandlerOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJSON"
                        }
                    }
                }
            }
        },
        "/ledger/transactions": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ledger"
                ],
                "summary": "Add transaction",
                "parameters": [
                    {
                        "description": "JSON",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ledger.TransactionAddHandlerInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ledger.TransactionAddHandlerOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJSON"
                        }
                    }
                }
            }
        },
        "/ledger/transactions/export": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "ledger"
                ],
                "summary": "Export list transactions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по дате ОТ в формате 2025-01-30 (год-месяц-день)",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по дате ДО в формате 2025-01-30 (год-месяц-день)",
                        "name": "date_to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CSV file"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJSON"
                        }
                    }
                }
            }
        },
        "/ledger/transactions/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "ledger"
                ],
                "summary": "Delete transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorJSON"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "ledger.BudgetAddHandlerInputPeriod": {
            "type": "object",
            "properties": {
                "month": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "ledger.BudgetGetHandlerOutput": {
            "type": "object",
            "properties": {
                "hitCache": {
                    "type": "boolean"
                },
                "item": {
                    "$ref": "#/definitions/ledger.BudgetOutput"
                }
            }
        },
...
        "ledger.BudgetOutput": {
            "type": "object",
            "properties": {
                "accountID": {
                    "type": "string"
                },
                "amount": {
                    "type": "string"
                },
                "categoryID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "period": {
                    "$ref": "#/definitions/ledger.BudgetOutputPeriod"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "ledger.BudgetOutputPeriod": {
            "type": "object",
            "properties": {
                "month": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "ledger.BudgetPatchHandlerInput": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "categoryID": {
                    "type": "integer"
                },
                "period": {
                    "$ref": "#/definitions/ledger.BudgetAddHandlerInputPeriod"
                }
            }
        },
        "ledger.BudgetPatchHandlerOutput": {
            "type": "object",
            "properties": {
                "item": {
                    "$ref": "#/definitions/ledger.BudgetOutput"
                }
            }
        },
...
        "ledger.ReportListHandlerOutput": {
            "type": "object",
            "properties": {
                "hitCache": {
                    "type": "boolean"
                },
                "reports": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ledger.ReportOutput"
                    }
                }
            }
        },
        "ledger.ReportOutput": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ledger.ReportOutputItem"
                    }
                },
                "periodEnd": {
                    "type": "string"
                },
                "periodStart": {
                    "type": "string"
                }
            }
        },
        "ledger.ReportOutputItem": {
            "type": "object",
            "properties": {
                "categoryID": {
                    "type": "integer"
                },
                "itemBudget": {
                    "type": "string"
                },
                "spentBudget": {
                    "type": "string"
                },
                "sum": {
                    "type": "string"
                }
            }
        },
        "ledger.TransactionAddHandlerInput": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "categoryID": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "isIncome": {
                    "type": "boolean"
                },
                "occurredOn": {
                    "type": "string"
                }
            }
        },
        "ledger.TransactionAddHandlerOutput": {
            "type": "object",
            "properties": {
                "item": {
                    "$ref": "#/definitions/ledger.TransactionOutput"
                }
            }
        },
        "ledger.TransactionOutput": {
            "type": "object",
            "properties": {
                "accountID": {
                    "type": "string"
                },
                "amount": {
                    "type": "string"
                },
                "categoryID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isIncome": {
                    "type": "boolean"
                },
                "occurredOn": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "middleware.ErrorJSON": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "hints": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "textCode": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}